Additional project restructuring 🚚
