+++ package.json
@@ -6,8 +6,8 @@
   "author": "Tyler Crosse",
   "license": "MIT",
   "scripts": {
-    "start": "cross-env NODE_ENV=production node compiled/server.js",
-    "dev": "cross-env NODE_ENV=development nodemon ./src/server/",
+    "start": "cross-env NODE_ENV=production node compiled/server/server.js",
+    "dev": "cross-env NODE_ENV=development nodemon --ignore client/ ./src/server/",
     "build": "rimraf ./compiled && npm run build:server && npm run build:client",
     "build:client": "cross-env NODE_ENV=production webpack --progress --profile --colors",
     "build:server": "cross-env NODE_ENV=production npm run babel -- ./src -d ./compiled",
@@ -32,7 +32,7 @@
     "cookie-parser": "^1.4.3",
     "cross-env": "^4.0.0",
     "css-loader": "^0.28.0",
-    "d3": "^4.2.7",
+    "d3": "^4.8.0",
     "eventemitter3": "^2.0.2",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
