+++ src/client/components/App.jsx
@@ -2,7 +2,7 @@ import React                  from 'react';
 import ReactDOM               from 'react-dom';
 import { connect }            from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as Actions           from './actions';
+import * as Actions           from '../actions';
 import Login                  from './user/Login.jsx';
 import LoggedIn               from './user/LoggedIn.jsx';
 import SearchContainer        from './search/SearchContainer.jsx';
