+++ src/client/components/search/Result.jsx
