+++ src/client/components/search/SearchContainer.jsx
@@ -1,10 +1,10 @@
 import React                  from 'react';
 import { connect }            from 'react-redux';
 import { bindActionCreators } from 'redux';
-import Searchbar    from './Searchbar.jsx';
-import ResultsList  from './ResultsList.jsx';
-import * as Actions from '../actions/';
-import                   './search.scss';
+import Searchbar              from './Searchbar.jsx';
+import ResultsList            from './ResultsList.jsx';
+import * as Actions           from '../../actions/';
+import                             './search.scss';
 
 function SearchContainer(props) {
   return (
