+++ src/client/components/search/search.scss
@@ -1,4 +1,4 @@
-@import '../mixins/_variables.scss';
+@import '../_variables.scss';
 
 .search {
   button {
