+++ src/client/components/viz/NodeDetails.jsx
