+++ src/client/index.js
@@ -1,7 +1,7 @@
 import React          from 'react';
 import ReactDOM       from 'react-dom';
 import { Provider }   from 'react-redux';
-import App            from './App';
+import App            from './components/App';
 import configureStore from './store/configureStore';
 
 const store = configureStore();
