--- webpack.config.prod.js
@@ -1,68 +1,0 @@
-
-
-const path = require('path');
-const webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const StatsPlugin = require('stats-webpack-plugin');
-const autoprefixer = require('autoprefixer');
-const precss = require('precss');
-
-module.exports = {
-  entry: [
-    path.join(__dirname, 'app/main.js')
-  ],
-  output: {
-    path: path.join(__dirname, '/dist/'),
-    filename: '[name]-[hash].min.js',
-    publicPath: '/'
-  },
-  plugins: [
-    new webpack.optimize.OccurenceOrderPlugin(),
-    new HtmlWebpackPlugin({
-      template: 'app/index.tpl.html',
-      inject: 'body',
-      filename: 'index.html'
-    }),
-    new ExtractTextPlugin('[name]-[hash].min.css'),
-    new webpack.optimize.UglifyJsPlugin({
-      compressor: {
-        warnings: false,
-        screw_ie8: true
-      }
-    }),
-    new StatsPlugin('webpack.stats.json', {
-      source: false,
-      modules: false
-    }),
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-    })
-  ],
-  module: {
-    loaders: [{
-      test: /\.jsx?$/,
-      exclude: /node_modules/,
-      loader: 'babel',
-      query: {
-        presets: ['es2015', 'stage-0', 'react']
-      }
-    }, {
-      test: /\.json?$/,
-      loader: 'json'
-    }, {
-      test: /\.scss$/,
-      loader: ExtractTextPlugin.extract('css-loader?modules&importLoaders=1&localIdentName=[path]___[name]__[local]___[hash:base64:5]!sass-loader')
-    }]
-  },
-  postcss() {
-    return [autoprefixer, precss];
-  },
-  sassLoader: {
-    includePaths: [path.join(__dirname, 'scss')]
-  }
-};
-
-// loader: ExtractTextPlugin.extract('style', 'css?modules&localIdentName=[name]---[local]---[hash:base64:5]!postcss')
-
-// https://github.com/postcss/postcss-loader/issues/42
